scorelib: drop tables in create if they exist
Old tables were kept in-memory when running unit tests with maven.
This caused tests to work in Eclipse, but fail on command line.

Changed create sql to drop tables (if they exists) before create,
and made db initialization to fail on error so the error of "old
data" would be seen also in Eclipse if it happens again.