scorelib: configure spring beans
Completed bean configuration class DbConfig and specificly its bean
jdbcTemplate. The bean is singleton, so within context same instance
is shared by clients. The bean instantion reads votings from input.txt
(a resource in classpath) and writes them into a tables that were
created by the config. It returns data source to the DB. As long as
in-memory database is used and as long as there is single context
in a jvm, this is fine.